IGVF-153-table-display
Changed the styling of the sortable table header cell so that the sorting icon takes only as much space as it needs in Safari and Firefox, as it already does in Chrome. But even with this fix, Safari displays a scroll bar even when not needed, because of the space the border takes. I fixed that by changing the border to an outline, but only on sortable tables. We might need to go that route for future uses of <DataGrid>.