IGVF-197-reload-redirect
Remove the <Session> redirect. I don’t know why I thought I needed that. For the most part it causes no harm, but if you reload a non-home page or directly enter a URL to a non-home page, it reloads to some other page, often but not always the home page.