IGVF-1617-collapsible-lists
Hopeful fix for the UI crash. The UI expected MultiplexedSample.construct_library_samples to be populated if any construct library samples exist in the `multiplexed_samples` array. But this appears not to be the case. Display the `@id` of the construct library if the corresponding construct library sample doesn’t exist in MultiplexedSample.construct_library_samples.